Feature:
Update of classes, (rental functional)
